refactor(worker-framework): extract generic worker management package
Move leader election, coordinator, dedicated-worker lifecycle, and generic
effect-rpc RpcPort helpers out of @dxos/client into @dxos/worker-framework,
with @dxos/client and @dxos/client-protocol rewired via injected callbacks.
Includes spec.md documenting the architecture and integration seams.
