+++ packages/sdk/client/src/services/dedicated/index.ts
@@ -2,10 +2,26 @@
 // Copyright 2026 DXOS.org
 //
 
-export * from './coordinator-worker';
-export * from './dedicated-worker-client-services';
-export * from './dedicated-worker';
-export * from './memory-coordinator';
-export * from './shared-worker-coordinator';
-export * from './single-client-coordinator';
-export * from './types';
+import {
+  MemoryWorkerCoordiantor,
+  SingleClientCoordinator,
+  createCoordinatorOnConnect,
+} from '@dxos/worker-framework/coordinator';
+export type {
+  DedicatedWorkerMessage,
+  WorkerCoordinator,
+  WorkerCoordinatorMessage,
+  WorkerOrPort,
+} from '@dxos/worker-framework/coordinator';
+
+export { MemoryWorkerCoordiantor, SingleClientCoordinator, createCoordinatorOnConnect };
+
+export { type RunDedicatedWorkerOptions, runDedicatedWorker } from './dedicated-worker';
+export {
+  type DedeciatedWorkerClientServicesOptions,
+  DedicatedWorkerClientServices,
+  LEADER_LOCK_KEY,
+  type LeaderTimeoutOptions,
+} from './dedicated-worker-client-services';
+
+export { SharedWorkerCoordinator } from './shared-worker-coordinator';
