+++ packages/sdk/worker-framework/moon.yml
@@ -1,0 +1,12 @@
+layer: library
+language: typescript
+tags:
+  - ts-build
+  - ts-test
+  - pack
+tasks:
+  compile:
+    args:
+      - '--entryPoint=src/index.ts'
+      - '--entryPoint=src/worker/index.ts'
+      - '--entryPoint=src/coordinator/index.ts'
