+++ packages/sdk/worker-framework/src/coordinator/index.ts
@@ -1,0 +1,14 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export * from './coordinator-worker';
+export * from './memory-coordinator';
+export * from './shared-worker-coordinator';
+export * from './single-client-coordinator';
+export type {
+  DedicatedWorkerMessage,
+  WorkerCoordinator,
+  WorkerCoordinatorMessage,
+  WorkerOrPort,
+} from '../internal/messages';
