+++ packages/sdk/worker-framework/src/index.ts
@@ -1,0 +1,13 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export * from './worker-connection';
+export { makeRpcClient, serveRpcGroup } from './internal/rpc';
+export type { RpcGroupServer, ServeRpcGroupOptions } from './internal/rpc';
+export type {
+  DedicatedWorkerMessage,
+  WorkerCoordinator,
+  WorkerCoordinatorMessage,
+  WorkerOrPort,
+} from './internal/messages';
