+++ packages/sdk/worker-framework/src/worker/index.ts
@@ -1,0 +1,6 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export * from './run-worker';
+export type { DedicatedWorkerMessage, WorkerEndpoint } from '../internal/messages';
