+++ release-please-config.json
@@ -1135,6 +1135,11 @@
         },
         {
           "type": "json",
+          "path": "packages/sdk/worker-framework/package.json",
+          "jsonpath": "$.version"
+        },
+        {
+          "type": "json",
           "path": "packages/stories/stories-assistant/package.json",
           "jsonpath": "$.version"
         },
