+++ tsconfig.all.json
@@ -664,6 +664,9 @@
       "path": "packages/sdk/types/tsconfig.json"
     },
     {
+      "path": "packages/sdk/worker-framework/tsconfig.json"
+    },
+    {
       "path": "packages/stories/stories-assistant/tsconfig.json"
     },
     {
